ChartScreen has been udpated
