+++ shared/src/commonMain/kotlin/ui/chart/ChartScreen.kt
@@ -203,53 +203,4 @@ class ChartScreen(
 
     }
 
-
-    @Composable
-    fun RenderBarChart(
-        title: String,
-        yAxisData: List<Double>,
-        xAxisData: List<String>,
-        color: Color
-    ) {
-        Box(Modifier.fillMaxWidth().height(400.dp)) {
-            BarChart(
-                chartParameters = listOf(
-                    BarParameters(
-                        title,
-                        data = yAxisData,
-                        barColor = color
-                    )
-                ),
-                gridColor = Color.DarkGray,
-                xAxisData = xAxisData,
-                isShowGrid = true,
-                animateChart = true,
-                showGridWithSpacer = true,
-                yAxisStyle = TextStyle(
-                    fontSize = 14.sp,
-                    color = Color.DarkGray,
-                ),
-                xAxisStyle = TextStyle(
-                    fontSize = 14.sp,
-                    color = Color.DarkGray,
-                    fontWeight = FontWeight.W400
-                ),
-                yAxisRange = 5,
-                barWidth = 20.dp,
-                spaceBetweenBars = Dimens.doubleSpace,
-                spaceBetweenGroups = Dimens.space,
-            )
-        }
-    }
-
-
-    @Composable
-    fun RenderTitle(title: String) {
-        Text(
-            text = title,
-            modifier = Modifier.padding(horizontal = Dimens.doubleSpace),
-            style = MaterialTheme.typography.titleLarge,
-            fontWeight = FontWeight.ExtraBold
-        )
-    }
 }
